+++ source/index.ts
@@ -1,2 +1,3 @@
+export * from './on-context-invalidated.js';
 export * from './on-extension-start.js';
 export * from './one-event.js';
